Switched out some of the meshes.
One of the issues with the original Sun Temple content is that it doesn't reuse meshes. Becuase of this, we have multiple meshes that are identical to eachother eating up resources. I looked though a few of the items in the Sun Temple prefab, and switched them to use the same mesh. This will reduce the amount of system memory used, the amount of video memory used, and possibly cut down the number of batch calls when rendering. The resources this frees up might be enough to allow nicer settings when I get around to baking the lighting.