Cut down number of identical meshes.
There are lots and lots of meshes that are identical, so getting rid of them all will take a bit of effort. By my math, I've reduced the amount of GPU memory use by over 4MB in this time. I know that isn't much, but some graphics cards don't have much memory to begin with. I've also enabled GPU Instancing in most of the materials. Which is needed for to cut down the number of draw calls when dealing with a non-static item collection that re-uses meshes.